requirements.txt
backports.cached-property;python_version<"3.8"
backports.zoneinfo;python_version<"3.9"
scramp>=1.2.2
